+++ src/ingest/auth/refresh/post.js
@@ -28,8 +28,8 @@ module.exports = async (req, res) => {
 
 		// return ok
 		return response.ok(req, res, {
-			expiresIn: parseInt(login.login.expires_in),
-			expires: moment().add(parseInt(login.login.expires_in), 's').toISOString(),
+			expiresIn: parseFloat(login.login.expires_in),
+			expires: moment().add(parseFloat(login.login.expires_in), 's').toISOString(),
 			
 			token: login.login.id_token,
 			refreshToken: login.login.refresh_token,
